refactor(sync): infra package + propagate Store5 sync errors
Organize the sync/ module like core/data's infra pattern — public API stays at
kpt.sync, plumbing moves to kpt.sync.infra:

- kpt.sync (consumer-facing): WorkScheduler, SyncTypes (WorkHandle/WorkMode/
  WorkStatus/NotificationContent), di.SyncModule
- kpt.sync.infra (plumbing): DataSyncWorker, NotificationWorker,
  DefaultWorkScheduler, SyncNotifier (+ android/ios/desktop/wasmJs actuals)

Consumers now see a clean kpt.sync.* surface; the workers/scheduler/notifier
internals are tucked under .infra. WorkerDeclarations + KoinExt import updated.

Fix Store5 error propagation in syncWith (CurrencyRepositoryImpl +
MacroIndicatorsRepositoryImpl): the terminal StoreReadResponse.Error was
collected then discarded, so a failed fetch still let snapshotSync stamp
success and skip the worker retry. Chain .requireData() so an Error throws →
snapshotSync propagates → DataSyncWorker returns WorkResult.retry().
(Addresses the CodeRabbit review on both impls.)

Store5 refresh stays in the repository (each repo owns its store/keys);
DataSyncWorker remains a thin fan-out orchestrator — the clean separation.

Remove the cmp-android build comment explaining the absent worker dep.
