+++ sync/src/androidMain/kotlin/kpt/sync/infra/SyncNotifier.android.kt
@@ -7,7 +7,7 @@
  *
  * See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
  */
-package kpt.sync
+package kpt.sync.infra
 
 import com.mmk.kmpnotifier.notification.configuration.NotificationPlatformConfiguration
