+++ sync/src/commonMain/kotlin/kpt/sync/infra/DefaultWorkScheduler.kt
@@ -12,7 +12,7 @@
     kotlin.uuid.ExperimentalUuidApi::class,
 )
 
-package kpt.sync
+package kpt.sync.infra
 
 import io.github.mobilebytelabs.worker.ExistingWorkPolicy
 import io.github.mobilebytelabs.worker.OutOfQuotaPolicy
@@ -23,6 +23,13 @@ import io.github.mobilebytelabs.worker.workDataOf
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import kotlin.time.Duration
+import kpt.sync.DATA_SYNC_WORK_NAME
+import kpt.sync.NOTIFICATION_WORK_NAME_PREFIX
+import kpt.sync.NotificationContent
+import kpt.sync.WorkHandle
+import kpt.sync.WorkMode
+import kpt.sync.WorkScheduler
+import kpt.sync.WorkStatus
 
 /**
  * Default [WorkScheduler] — thin façade over worker-kmp's [WorkManager].
